上传了小程序，更换了stomp+uni.websocket 实现客户端消息发送和接收

+++ Java/src/main/java/com/jxufe_yzt/java/service/UserService.java
@@ -59,13 +59,13 @@ public class UserService {
     public void delete (User user) {
         userMapper.delete(user);
     }
-    public void delete(Integer userId) {
+    public void delete(Long userId) {
         userMapper.deleteByPrimaryKey(userId);
     }
     public void update(User user) {
         userMapper.updateByPrimaryKeySelective(user);
     }
-    public User selectById(Integer userId) {
+    public User selectById(Long userId) {
         return userMapper.selectByPrimaryKey(userId);
     }
 }
